+++ src/main/java/com/sp/util/keyframes/Keyframe.java
@@ -3,7 +3,7 @@ package com.sp.util.keyframes;
 
 public class Keyframe {
     private boolean initialized;
-    private final float keyframeTime;
+    private final double keyframeTime;
     private final Runnable initAction;
     private final KeyframeAction action;
 
@@ -11,7 +11,7 @@ public class Keyframe {
      * No actions taken. Meant to act as a delay
      * @param keyframeTime The point in time (0 -> 1) that the keyframe should be placed
      */
-    public Keyframe(float keyframeTime) {
+    public Keyframe(double keyframeTime) {
         this(keyframeTime, () -> {}, (globalTime, localTime) -> {});
     }
 
@@ -20,7 +20,7 @@ public class Keyframe {
      * @param keyframeTime The point in time (0 -> 1) that the keyframe should be placed
      * @param action The action that should be called every time the keyframe is updated
      */
-    public Keyframe(float keyframeTime, KeyframeAction action) {
+    public Keyframe(double keyframeTime, KeyframeAction action) {
         this(keyframeTime, () -> {}, action);
     }
 
@@ -29,7 +29,7 @@ public class Keyframe {
      * @param keyframeTime The point in time (0 -> 1) that the keyframe should be placed
      * @param initAction The action that should be called once at the beginning of the keyframe
      */
-    public Keyframe(float keyframeTime, Runnable initAction) {
+    public Keyframe(double keyframeTime, Runnable initAction) {
         this(keyframeTime, initAction, (globalTime, localTime) -> {});
     }
 
@@ -39,7 +39,7 @@ public class Keyframe {
      * @param initAction The action that should be called once at the beginning of the keyframe
      * @param action The action that should be called every time the keyframe is updated
      */
-    public Keyframe(float keyframeTime, Runnable initAction, KeyframeAction action) {
+    public Keyframe(double keyframeTime, Runnable initAction, KeyframeAction action) {
         this.keyframeTime = keyframeTime;
         this.initAction = initAction;
         this.action = action;
@@ -53,7 +53,7 @@ public class Keyframe {
         return this.initAction;
     }
 
-    protected float getKeyframeTime() {
+    protected double getKeyframeTime() {
         return this.keyframeTime;
     }
 
@@ -73,6 +73,6 @@ public class Keyframe {
          * @param globalTime The time (from 0 -> 1) from the start of the animation to the end
          * @param localTime  The time (from 0 -> 1) from the start of the keyframe to the start of the next keyframe
          */
-        void run(float globalTime, float localTime);
+        void run(double globalTime, double localTime);
     }
 }
